Workaround for the MacOS nightly builds issue
The MacOS nightly build seems to always crash on
xbmc.Player().play(url). However, it seems to work by directly
processing the link in the episodes list item. This complicates things
with the subtitles, on the other hand. This is a quick fix for the MacOS
issue. Hopefully, this XBMC bug will be sorted out soon.